+++ include/liboffsetfinder64/common.h
