+++ include/liboffsetfinder64/exception.hpp
